backend/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
pydantic[email]==2.5.0